requirements.txt
requests
colorama
dnspython
graphviz
python-whois
selenium
